refactor(cli): stage 1 — extract shared CLI helpers into _common.py
Move TalkativeParser, check, and required_length out of
mk_prepare_receptor.py into meeko/cli/_common.py. TalkativeParser is
now a factory (make_talkative_parser) parametrized by the calling
script's path so the error message stays accurate.

The receptor module re-exposes TalkativeParser as a module attribute
(via make_talkative_parser(path_to_this_script)) so any
``from meeko.cli.mk_prepare_receptor import TalkativeParser`` keeps
working. check and required_length are imported directly.

ligand CLI can now reuse these helpers without copy-paste (next stages).

92 tests pass, 4 skipped.
